fix(web): close #1720 #1462
