+++ packages/remax/web.d.ts
@@ -1,0 +1,1 @@
+export * from '@remax/web';
