+++ packages/remax/web.js
@@ -1,0 +1,1 @@
+export * from '@remax/web';
